adding logic, fixing

+++ src/repository/contacts.js
@@ -6,39 +6,58 @@ class ContactsRepository {
     this.model = Contact
   }
 
-  async listContacts() {
-    const results = await this.model.find({})
-    return results
-  }
-
-  async getContactById(id) {
-    const result = await this.model.findOne({ _id: id })
+  async listContacts(userId, { page = 1, limit = 5, sortBy, sortByDesc, filter, favorite = null }) {
+    const optionSearch = { owner: userId }
+    if (favorite !== null) {
+      optionSearch.favorite = favorite
+    }
+    const result = await this.model.paginate(optionSearch, {
+      limit,
+      page,
+      sort: {
+        ...(sortBy ? { [`${sortBy}`]: 1 } : {}),
+        ...(sortByDesc ? { [`${sortByDesc}`]: -1 } : {})
+      },
+      select: filter ? filter.split('|').join(' ') : '',
+      populate: {
+        path: 'owner',
+        select: 'name email subscription',
+      },
+    })
     return result
   }
 
-  async addContact(body) {
-    const newContact = await this.model.create(body)
+  async getContactById(userId, id) {
+    const result = await this.model.findOne({ _id: id, owner: userId }).populate({
+      path: 'owner',
+      select: 'name email subscription',
+    })
+    return result
+  }
+
+  async addContact(userId, body) {
+    const newContact = await this.model.create({ ...body, owner: userId })
 
     return newContact
   }
 
-  async updateContact(id, body) {
+  async updateContact(userId, id, body) {
     const updatedClient = await this.model.findByIdAndUpdate(
-      { _id: id },
+      { _id: id, owner: userId },
       { ...body },
       { new: true }
     )
     return updatedClient
   }
 
-  async removeContact(id) {
-    const removedClient = await this.model.findByIdAndRemove({ _id: id },
+  async removeContact(userId, id) {
+    const removedClient = await this.model.findByIdAndRemove({ _id: id, owner: userId },
     )
     return removedClient
   }
 
-  async updateStatusContact(id, body) {
-    const updatedClient = await this.model.findByIdAndUpdate({ _id: id },
+  async updateStatusContact(userId, id, body) {
+    const updatedClient = await this.model.findByIdAndUpdate({ _id: id, owner: userId },
       { ...body },
       { new: true }
     )
